Move draw code to DrawViewController
